move util package to internal/util
