+++ internal/service/handler_delete_test.go
@@ -22,7 +22,7 @@ import (
 	"testing"
 
 	"github.com/CrunchyData/pg_featureserv/internal/data"
-	"github.com/CrunchyData/pg_featureserv/util"
+	"github.com/CrunchyData/pg_featureserv/internal/util"
 	"github.com/getkin/kin-openapi/openapi3"
 )
