+++ internal/service/handler_update_test.go
@@ -21,7 +21,7 @@ import (
 	"testing"
 
 	"github.com/CrunchyData/pg_featureserv/internal/api"
-	"github.com/CrunchyData/pg_featureserv/util"
+	"github.com/CrunchyData/pg_featureserv/internal/util"
 	"github.com/getkin/kin-openapi/openapi3"
 )
